move grundeigentumArtMapping to shared location, rename shared eigentum helpers, begin filling out fillGrundeigentum function
