+++ app/flows/shared/finanzielleAngaben/context.ts
@@ -18,6 +18,8 @@ export const Eigentuemer = z.enum(
   customRequiredErrorMessage,
 );
 
+export type Eigentumer = z.infer<typeof Eigentuemer>;
+
 const MINUS_150_YEARS = -150;
 
 export const GrundeigentumArt = z.enum(
