+++ app/services/pdf/prozesskostenhilfe/I_zahlungsverpflichtungen.ts
@@ -3,8 +3,8 @@ import type { ProzesskostenhilfeFormularContext } from "~/flows/prozesskostenhil
 import { finanzielleAngabeEinkuenfteGuards as einkuenfteGuards } from "~/flows/prozesskostenhilfeFormular/finanzielleAngaben/einkuenfte/guards";
 import type { PkhPdfFillFunction } from ".";
 import type { AttachmentEntries } from "../attachment";
-import { eigentuemerMapping } from "../shared/eigentuemerMapping";
 import { SEE_IN_ATTACHMENT_DESCRIPTION } from "../beratungshilfe/sections/E_unterhalt";
+import { eigentuemerMapping } from "../shared/eigentumHelpers";
 
 type Zahlung = NonNullable<
   | ProzesskostenhilfeFormularContext["ratenzahlungen"]
